Move buffer pack in one to separate file
