+++ src/bvals/cc/bvals_cc_in_one.hpp
@@ -1,0 +1,36 @@
+//========================================================================================
+// Parthenon performance portable AMR framework
+// Copyright(C) 2020 The Parthenon collaboration
+// Licensed under the 3-clause BSD License, see LICENSE file for details
+//========================================================================================
+// (C) (or copyright) 2020. Triad National Security, LLC. All rights reserved.
+//
+// This program was produced under U.S. Government contract 89233218CNA000001 for Los
+// Alamos National Laboratory (LANL), which is operated by Triad National Security, LLC
+// for the U.S. Department of Energy/National Nuclear Security Administration. All rights
+// in the program are reserved by Triad National Security, LLC, and the U.S. Department
+// of Energy/National Nuclear Security Administration. The Government is granted for
+// itself and others acting on its behalf a nonexclusive, paid-up, irrevocable worldwide
+// license in this material to reproduce, prepare derivative works, distribute copies to
+// the public, perform publicly and display publicly, and to permit others to do so.
+//========================================================================================
+
+#ifndef BVALS_CC_BVALS_CC_IN_ONE_HPP_
+#define BVALS_CC_BVALS_CC_IN_ONE_HPP_
+
+#include <string>
+
+#include "basic_types.hpp"
+#include "mesh/mesh.hpp"
+
+namespace parthenon {
+namespace cell_centered_bvars {
+auto SendBoundaryBuffers(BlockList_t &blocks, const std::string &container_name)
+    -> TaskStatus;
+auto ReceiveBoundaryBuffers(BlockList_t &blocks, const std::string &container_name)
+    -> TaskStatus;
+auto SetBoundaries(BlockList_t &blocks, const std::string &container_name) -> TaskStatus;
+} // namespace cell_centered_bvars
+} // namespace parthenon
+
+#endif // BVALS_CC_BVALS_CC_IN_ONE_HPP_
